further split between methods and config
