+++ python/semantic_kernel/connectors/ai/open_ai/services/azure_text_completion.py
@@ -2,29 +2,20 @@
 
 
 from logging import Logger
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, Optional
 
-from pydantic import Field, constr
-
-from semantic_kernel.connectors.ai import TextCompletionClientBase
-from semantic_kernel.connectors.ai.complete_request_settings import (
-    CompleteRequestSettings,
+from semantic_kernel.connectors.ai.open_ai.services.base_config_azure import (
+    BaseAzureConfig,
 )
-from semantic_kernel.connectors.ai.open_ai.services.base_open_ai_service_calls import (
+from semantic_kernel.connectors.ai.open_ai.services.base_open_ai_functions import (
     OpenAIModelTypes,
-    OpenAIServiceCalls,
 )
-from semantic_kernel.sk_pydantic import HttpsUrl
+from semantic_kernel.connectors.ai.open_ai.services.base_text_completion import (
+    BaseTextCompletion,
+)
 
 
-class AzureTextCompletion(OpenAIServiceCalls, TextCompletionClientBase):
-    model_id: constr(strip_whitespace=True, min_length=1) = Field(
-        ..., alias="deployment_name"
-    )
-    api_version: Optional[str] = None
-    endpoint: Optional[HttpsUrl] = None
-    api_type: str = "azure"
-
+class AzureTextCompletion(BaseAzureConfig, BaseTextCompletion):
     def __init__(
         self,
         deployment_name: str,
@@ -60,19 +51,19 @@ class AzureTextCompletion(OpenAIServiceCalls, TextCompletionClientBase):
         if logger:
             logger.warning("The 'logger' argument is deprecated, use 'log' instead.")
         super().__init__(
-            model_id=deployment_name,
+            deployment_name=deployment_name,
             endpoint=endpoint,
             api_key=api_key,
             api_version=api_version,
             log=log or logger,
             model_type=OpenAIModelTypes.TEXT,
-            api_type="azure_ad" if ad_auth else "azure",
+            ad_auth=ad_auth,
         )
 
     @classmethod
     def from_dict(cls, settings: Dict[str, str]) -> "AzureTextCompletion":
         """
-        Initialize an AzureChatCompletion service from a dictionary of settings.
+        Initialize an Azure OpenAI service from a dictionary of settings.
 
         Arguments:
             settings: A dictionary of settings for the service.
@@ -91,59 +82,3 @@ class AzureTextCompletion(OpenAIServiceCalls, TextCompletionClientBase):
             ad_auth=settings.get("ad_auth", False),
             log=settings.get("log"),
         )
-
-    def to_dict(self) -> Dict[str, str]:
-        return self.dict(
-            exclude={
-                "prompt_tokens",
-                "completion_tokens",
-                "total_tokens",
-                "api_type",
-                "org_id",
-                "model_type",
-            },
-            by_alias=True,
-            exclude_none=True,
-        )
-
-    def get_model_args(self) -> Dict[str, Any]:
-        return {
-            "deployment_id": self.model_id,
-            "api_type": self.api_type,
-            "api_base": self.endpoint,
-            "api_version": self.api_version,
-        }
-
-    async def complete_stream_async(
-        self,
-        prompt: str,
-        settings: CompleteRequestSettings,
-        logger: Optional[Logger] = None,
-    ):
-        response = await self._send_request(
-            prompt=prompt, request_settings=settings, stream=True
-        )
-
-        async for chunk in response:
-            if settings.number_of_responses > 1:
-                for choice in chunk.choices:
-                    completions = [""] * settings.number_of_responses
-                    completions[choice.index] = choice.text
-                    yield completions
-            else:
-                yield chunk.choices[0].text
-
-    async def complete_async(
-        self,
-        prompt: str,
-        settings: CompleteRequestSettings,
-        logger: Optional[Logger] = None,
-    ) -> Union[str, List[str]]:
-        response = await self._send_request(
-            prompt=prompt, request_settings=settings, stream=False
-        )
-
-        if len(response.choices) == 1:
-            return response.choices[0].text
-        else:
-            return [choice.text for choice in response.choices]
