+++ python/semantic_kernel/connectors/ai/open_ai/services/base_config_azure.py
@@ -2,35 +2,40 @@
 
 
 from logging import Logger
-from typing import Dict, Optional
+from typing import TYPE_CHECKING, Any, Dict, Optional
 
-from semantic_kernel.connectors.ai.open_ai.services.base_chat_completions import (
-    BaseChatCompletion,
-)
-from semantic_kernel.connectors.ai.open_ai.services.base_config_azure import (
-    BaseAzureConfig,
-)
+from pydantic import Field, constr
+
 from semantic_kernel.connectors.ai.open_ai.services.base_open_ai_functions import (
     OpenAIModelTypes,
+    OpenAIServiceCalls,
 )
-from semantic_kernel.connectors.ai.open_ai.services.base_text_completion import (
-    BaseTextCompletion,
-)
+from semantic_kernel.sk_pydantic import HttpsUrl
+
+if TYPE_CHECKING:
+    pass
 
 
-class AzureChatCompletion(BaseAzureConfig, BaseChatCompletion, BaseTextCompletion):
+class BaseAzureConfig(OpenAIServiceCalls):
+    model_id: constr(strip_whitespace=True, min_length=1) = Field(
+        ..., alias="deployment_name"
+    )
+    api_version: Optional[str] = None
+    endpoint: Optional[HttpsUrl] = None
+    api_type: str = "azure"
+
     def __init__(
         self,
         deployment_name: str,
         endpoint: str,
         api_key: str,
+        model_type: Optional[OpenAIModelTypes],
         api_version: str = "2022-12-01",
-        ad_auth=False,
+        ad_auth: bool = False,
         log: Optional[Logger] = None,
-        logger: Optional[Logger] = None,
     ) -> None:
         """
-        Initialize an AzureChatCompletion service.
+        Initialize an AzureTextCompletion service.
 
         Arguments:
             deployment_name: The name of the Azure deployment. This value
@@ -51,37 +56,34 @@ class AzureChatCompletion(BaseAzureConfig, BaseChatCompletion, BaseTextCompletio
             log: The logger instance to use. (Optional)
             logger: deprecated, use 'log' instead.
         """
-        if logger:
-            logger.warning("The 'logger' argument is deprecated, use 'log' instead.")
         super().__init__(
-            deployment_name=deployment_name,
+            model_id=deployment_name,
             endpoint=endpoint,
             api_key=api_key,
-            model_type=OpenAIModelTypes.CHAT,
             api_version=api_version,
-            log=log or logger,
-            ad_auth=ad_auth,
+            log=log,
+            model_type=model_type,
+            api_type="azure_ad" if ad_auth else "azure",
         )
 
-    @classmethod
-    def from_dict(cls, settings: Dict[str, str]) -> "AzureChatCompletion":
-        """
-        Initialize an Azure OpenAI service from a dictionary of settings.
-
-        Arguments:
-            settings: A dictionary of settings for the service.
-                should contains keys: deployment_name, endpoint, api_key
-                and optionally: api_version, ad_auth, log
-        """
-        if "api_type" in settings:
-            settings["ad_auth"] = settings["api_type"] == "azure_ad"
-            del settings["api_type"]
-
-        return AzureChatCompletion(
-            deployment_name=settings["deployment_name"],
-            endpoint=settings["endpoint"],
-            api_key=settings["api_key"],
-            api_version=settings.get("api_version"),
-            ad_auth=settings.get("ad_auth", False),
-            log=settings.get("log"),
+    def to_dict(self) -> Dict[str, str]:
+        return self.dict(
+            exclude={
+                "prompt_tokens",
+                "completion_tokens",
+                "total_tokens",
+                "api_type",
+                "org_id",
+                "model_type",
+            },
+            by_alias=True,
+            exclude_none=True,
         )
+
+    def get_model_args(self) -> Dict[str, Any]:
+        return {
+            "deployment_id": self.model_id,
+            "api_type": self.api_type,
+            "api_base": self.endpoint,
+            "api_version": self.api_version,
+        }
