+++ python/semantic_kernel/connectors/ai/open_ai/services/base_model_types.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class OpenAIModelTypes(Enum):
+    TEXT = "text"
+    CHAT = "chat"
+    EMBEDDING = "embedding"
