+++ python/tests/unit/ai/open_ai/services/test_azure_text_completion.py
@@ -109,7 +109,7 @@ def test_azure_text_completion_init_with_invalid_endpoint() -> None:
 async def test_azure_text_completion_call_with_parameters() -> None:
     mock_openai = AsyncMock()
     with patch(
-        "semantic_kernel.connectors.ai.open_ai.services.base_open_ai_service_calls.openai",
+        "semantic_kernel.connectors.ai.open_ai.services.base_open_ai_functions.openai",
         new=mock_openai,
     ):
         deployment_name = "test_deployment"
@@ -154,7 +154,7 @@ async def test_azure_text_completion_call_with_parameters() -> None:
 async def test_azure_text_completion_call_with_parameters_logit_bias_not_none() -> None:
     mock_openai = AsyncMock()
     with patch(
-        "semantic_kernel.connectors.ai.open_ai.services.base_open_ai_service_calls.openai",
+        "semantic_kernel.connectors.ai.open_ai.services.base_open_ai_functions.openai",
         new=mock_openai,
     ):
         deployment_name = "test_deployment"
